streamlit_app/streamlit_app/requirements.txt
# Streamlit App Dependencies

# Core
streamlit>=1.28.0

# PyTorch (CUDA 11.8)
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face
huggingface-hub>=0.19.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
